% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_sample.R
\name{geom_sf_sample}
\alias{geom_sf_sample}
\title{Visualise Sf Objectjects with Uncertainty}
\usage{
geom_sf_sample(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "sample",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n = NULL,
  ...
)
}
\arguments{
\item{n}{A parameter used to control the number of cells in each grid. Each area is broken up into an nxn grid}

\item{fill}{A variable of distributions from the `distributional` package}
}
\value{
A ggplot2 geom representing a sf_sample which can be added to a ggplot object
}
\description{
Identical to geom_sf, except that the fill for each area will be a distribution. 
This function will replace the fill area with a grid, where each cell is filled with an outcome from the fill distribution.
}
\examples{
# In it's most basic form, the geom will make a subdivision 
library(ggplot2)
toymap |>  
  ggplot() + 
  geom_sf_sample(aes(geometry = geometry, fill=temp_dist))
# The original borders of the sf object can be hard to see, 
 # so layering the original geometry on top can help to see the original boundaries
toymap |>  
  ggplot() + 
  geom_sf_sample(aes(geometry = geometry, fill=temp_dist), linewidth=0.1, n=5) + 
  geom_sf(aes(geometry=geometry), fill=NA, linewidth=1)
}
