% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-violin-sample.R, R/stat-ydensity-sample.R
\docType{data}
\name{geom_violin_sample}
\alias{geom_violin_sample}
\alias{StatYdensitySample}
\alias{stat_ydensity_sample}
\title{Violin plots with uncertainty}
\usage{
geom_violin_sample(
  mapping = NULL,
  data = NULL,
  stat = "ydensity_sample",
  position = "identity",
  ...,
  times = 10,
  quantile_gp = list(linetype = 0),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_ydensity_sample(
  mapping = NULL,
  data = NULL,
  geom = "violin",
  position = "identity",
  ...,
  times = 10,
  width = NULL,
  bw = "nrd0",
  adjust = 1,
  kernel = "gaussian",
  trim = TRUE,
  scale = "area",
  drop = TRUE,
  bounds = c(-Inf, Inf),
  quantiles = c(0.25, 0.5, 0.75),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{times}{A parameter used to control the number of values sampled from each distribution.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{geom, stat}{Use to override the default connection between
\code{geom_violin()} and \code{stat_ydensity()}. For more information about
overriding these connections, see how the \link[ggplot2:layer_stats]{stat} and
\link[ggplot2:layer_geoms]{geom} arguments work.}

\item{bw}{The smoothing bandwidth to be used.
If numeric, the standard deviation of the smoothing kernel.
If character, a rule to choose the bandwidth, as listed in
\code{\link[stats:bandwidth]{stats::bw.nrd()}}. Note that automatic calculation of the bandwidth does
not take weights into account.}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, \code{adjust = 1/2} means use half of the default bandwidth.}

\item{kernel}{Kernel. See list of available kernels in \code{\link[=density]{density()}}.}

\item{trim}{If \code{TRUE} (default), trim the tails of the violins
to the range of the data. If \code{FALSE}, don't trim the tails.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}

\item{drop}{Whether to discard groups with less than 2 observations
(\code{TRUE}, default) or keep such groups for position adjustment purposes
(\code{FALSE}).}

\item{bounds}{Known lower and upper bounds for estimated data. Default
\code{c(-Inf, Inf)} means that there are no (finite) bounds. If any bound is
finite, boundary effect of default density estimation will be corrected by
reflecting tails outside \code{bounds} around their closest edge. Data points
outside of bounds are removed with a warning.}

\item{quantiles}{If not \code{NULL} (default), compute the \code{quantile} variable
and draw horizontal lines at the given quantiles in \code{geom_violin()}.}
}
\description{
Identical to geom_violin, except that it will accept a distribution in
place of any of the usual aesthetics.
}
\examples{
library(ggplot2)
library(dplyr)
library(distributional)

# have to make factor variable, probably easier ways to do it
uncertain_mtcars2 <- uncertain_mtcars |>
  rowwise() |> #must have this or the distributions get mixed up
  mutate(cyl_factor = dist_sample(list(factor(unlist(generate(cyl,100))))))

# plot set up
p <- ggplot(mtcars, aes(factor(cyl), mpg))
q <- ggplot(uncertain_mtcars2, aes(cyl_factor, mpg))

# ggplot
p + geom_violin()
# ggdibbler
q + geom_violin_sample(alpha=0.1)

# Orientation follows the discrete axis
# ggplot
ggplot(mtcars, aes(mpg, factor(cyl))) +
  geom_violin()
# ggdibbler
ggplot(uncertain_mtcars2, aes(mpg, cyl_factor)) +
  geom_violin_sample(alpha=0.1)

# ggplot
p + geom_violin() + 
  geom_jitter(height = 0, width = 0.1)
# ggdibbler
q + geom_violin_sample(alpha=0.1) + 
  geom_jitter_sample(height = 0, width = 0.1, size=0.1)

# Scale maximum width proportional to sample size:
# ggplot
p + geom_violin(scale = "count")
# ggdibbler
q + geom_violin_sample(scale = "count")

# Scale maximum width to 1 for all violins:
# ggplot
p + geom_violin(scale = "width")
# ggdibbler
q + geom_violin_sample(scale = "width", alpha=0.1)

# Default is to trim violins to the range of the data. To disable:
# ggplot
p + geom_violin(trim = FALSE)
# ggdibbler
q + geom_violin_sample(trim = FALSE, alpha=0.1)

# Use a smaller bandwidth for closer density fit (default is 1).
# ggplot
p + geom_violin(adjust = .5)
# ggdibbler
q + geom_violin_sample(adjust = .5, alpha=0.1)


# Add aesthetic mappings

# ggplot
ggplot(mtcars, aes(factor(cyl), mpg)) + 
  geom_violin(aes(fill = cyl))
# ggdibbler
ggplot(uncertain_mtcars2, aes(cyl_factor, mpg)) + 
  geom_violin_sample(aes(fill = after_stat(x))) 

# ggplot
p + geom_violin(aes(fill = factor(cyl)))
# ggdibbler
q + geom_violin_sample(aes(fill = factor(after_stat(x))))

# Set aesthetics to fixed value
# ggplot
p + geom_violin(fill = "grey80", colour = "#3366FF")
# ggdibbler
q + geom_violin_sample(fill = "grey80", colour = "#3366FF", alpha=0.1)
}
\keyword{datasets}
