% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdibbler-package.R
\docType{package}
\name{ggdibbler-vctrs}
\alias{ggdibbler}
\alias{ggdibbler-package}
\alias{ggdibbler-vctrs}
\title{Internal vctrs methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package for visualising uncertainty with the goal of signal supression. Usually, uncertainty visualisation focuses on expressing uncertainty as a distribution or probability, whereas ggdibble differentiates itself by viewing an uncertainty visualisation as an existing graphic that has been adjusted to incorperate uncertainty in the estimates. You provide the code for an existing plot, but repalace one of the variables with a vector of distributions, and it will convert the visualisation into it's signal supression counterpart.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://harriet-mason.github.io/ggdibbler/}
}

}
\author{
\strong{Maintainer}: Harriet Mason \email{harriet.m.mason@gmail.com} (\href{https://orcid.org/0009-0007-4568-8215}{ORCID})

Authors:
\itemize{
  \item Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Sarah Goodwin \email{sarah.goodwin@monash.edu} (\href{https://orcid.org/0000-0001-8894-8282}{ORCID})
  \item Susan VanderPlas \email{susan.vanderplas@unl.edu} (\href{https://orcid.org/0000-0002-3803-0972}{ORCID})
}

}
\keyword{internal}
