% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsup.R
\name{pal_bivariate}
\alias{pal_bivariate}
\title{Generates a bivariate palette}
\usage{
pal_bivariate(colours, type = "desaturate", n = 4, amount = 1)
}
\arguments{
\item{colours}{character vector (of length d) of hex colours.}

\item{type}{a character to highlight how the values will be transformed between layers of the palette. Can be one of "desaturate", "lighten", or "transparency". Is set to "desaturate" by default.}

\item{n}{an integer representing the number of times the base colours should be transformed. I.e. the depth of the bivariate palette.}

\item{amount}{an integer representing the amount of desaturation, lightening, or transparency between the base colour and the final colour in the palette (with 1 being completely desaturated). Can also be a numeric vector of length n that provides the amount of desaturation occuring at each layer.}
}
\value{
A character vector representing the n*d colours in the bivariate palette
}
\description{
This function creates a bivariate palette, so that the previously 1D colour palette can become 2D, which allows variance to be mapped.
}
\examples{
colourvalues <- colorspace::sequential_hcl(4, palette = "YlOrRd")
pal_bivariate(colourvalues)

}
