% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-identity-dodge.R
\docType{data}
\name{position_identity_dodge}
\alias{position_identity_dodge}
\alias{PositionIdentityDodge}
\title{TITLE}
\usage{
position_identity_dodge(
  width = NULL,
  preserve = "single",
  orientation = "x",
  reverse = FALSE
)
}
\description{
DESCRIPTION
}
\examples{
print("x")
}
\keyword{datasets}
