% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-identity-nested.R
\docType{data}
\name{position_identity_identity}
\alias{position_identity_identity}
\alias{position_identity_dodge}
\alias{PositionIdentityDodge}
\alias{position_identity_stack}
\alias{PositionIdentityStack}
\title{Nested identity positions}
\usage{
position_identity_identity()

position_identity_dodge(
  width = NULL,
  preserve = "single",
  orientation = "x",
  reverse = FALSE
)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the \code{"total"} width of all elements
at a position, or the width of a \code{"single"} element?}

\item{orientation}{Fallback orientation when the layer or the data does not
indicate an explicit orientation, like \code{geom_point()}. Can be \code{"x"}
(default) or \code{"y"}.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}

\item{vjust}{Vertical adjustment for geoms that have a position
(like points or lines), not a dimension (like bars or areas). Set to
\code{0} to align with the bottom, \code{0.5} for the middle,
and \code{1} (the default) for the top.}
}
\description{
These functions use nested positioning for distributional data, where the
original plot has a identity position. This allows you to set different 
position adjustments for the "main" and "distribution" parts of your plot.
}
\examples{
# Standard ggplots often have a position adjustment to fix overplotting
# plot with overplotting
library(ggplot2)
ggplot(mpg, aes(class)) + 
  geom_bar(aes(fill = drv), alpha=0.5,
           position = "identity")

# sometimes ggdibbler functions call for more control over these 
# overplotting adjustments
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "identity", alpha=0.1)
# is the same as...
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "identity_identity", alpha=0.1)
  
# nested positions allows us to differentiate which postion adjustments
# are used for the plot groups vs the distribution samples
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "identity_dodge", alpha=0.7)
}
\keyword{datasets}
