% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-stack-dodge.R
\docType{data}
\name{position_stack_dodge}
\alias{position_stack_dodge}
\alias{PositionStackDodge}
\title{TITLE}
\usage{
position_stack_dodge(
  vjust = 1,
  reverse = FALSE,
  width = NULL,
  preserve = "single",
  orientation = "x"
)
}
\description{
DESCRIPTION
}
\examples{
print("x")
}
\keyword{datasets}
