% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-stack-identity.R
\docType{data}
\name{position_stack_identity}
\alias{position_stack_identity}
\alias{PositionStackIdentity}
\title{TITLE}
\usage{
position_stack_identity(vjust = 1, reverse = FALSE)
}
\description{
DESCRIPTION
}
\examples{
print("x")
}
\keyword{datasets}
