% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-sf-sample.R
\name{sample_subdivide_sf}
\alias{sample_subdivide_sf}
\title{Sample expand function for sf objects}
\usage{
sample_subdivide_sf(data, times)
}
\description{
Sample expand function for sf objects
}
\keyword{internal}
