% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-type.R
\name{scale_type.distribution}
\alias{scale_type.distribution}
\title{Sets default scale for distribution objects
Default scale for distribution objects is continuous}
\usage{
\method{scale_type}{distribution}(x)
}
\arguments{
\item{x}{variable being scaled}
}
\description{
Sets default scale for distribution objects
Default scale for distribution objects is continuous
}
