% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_identity_subdivide.R
\name{square_grid}
\alias{square_grid}
\title{Internal function for finding the "most square" factors of a number}
\usage{
square_grid(x)
}
\description{
Internal function for finding the "most square" factors of a number
}
\keyword{internal}
