% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_temp}
\alias{toy_temp}
\title{A toy data set that has the ambient temperature as measured by a collection of citizen scientists for each Iowa county}
\format{
A tibble with 99 observations and 4 variables
\describe{
  \item{county_name}{the name of each Iowa county}
  \item{recorded_temp}{the ambient temperature recorded by the citizen scientist}
  \item{scientistID}{the ID number for the scientist who made the recording}
  \item{county_geometry}{the shape file for each county of Iowa}
  \item{county_longitude}{the centroid longitude for each county of Iowa}
  \item{county_latitude}{the centroid latitude for each county of Iowa}
}
}
\description{
There are several measurements for each county, with no location marker
 for individual scientists to preserve anonyminity. Counties can have 
 different numbers of observations as well as a different levels of variance
 between the observations in the county.
}
