% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_temp_dist}
\alias{toy_temp_dist}
\title{A toy data set that provides data for a map with the temperature of each area represented by a random variable.}
\format{
A tibble with 99 observations and 4 variables
\describe{
  \item{county_name}{the name of each Iowa county}
  \item{temp_dist}{the temperature of each county as a distribution}
  \item{county_geometry}{the shape file for each county of Iowa}
}
}
\description{
The map shows a wave pattern in temperature on the state of Iowa.
Each estimate also has an uncertainty component added, and is represented as a distribution
}
