% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-distribution.R
\name{transform_distribution}
\alias{transform_distribution}
\title{Transformation for distributions (class Distribution)}
\usage{
transform_distribution()
}
\description{
Transformation for distributions (class Distribution)
}
\examples{
dists <- distributional::dist_normal(seq(1:10), 1)
t <- transform_distribution()
t$transform(dists)
t$inverse(t$transform(dists))
}
