% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uncertain_economics}
\alias{uncertain_economics}
\alias{uncertain_economics_long}
\title{An uncertain version of the economics data from`ggplot2`}
\format{
A data frame with almost 574 observations and 6 variables:
\describe{
  \item{date}{A deterministic variable - Month of data collection}
  \item{pce}{Normal random variable - personal consumption expenditures, in billions of dollars}
  \item{pop}{Normal random variable - total population, in thousands}
  \item{psavert}{Normal random variable - personal savings rate}
  \item{uempmed}{Normal random variable - median duration of unemployment, in weeks}
  \item{unemploy}{Normal random variable - number of unemployed in thousands}
  }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2870 rows and 4 columns.
}
\usage{
uncertain_economics_long
}
\description{
This dataset is identical to the economics data, except that every  
variable in the data set is represented by a normal
random variable. The original `economics` dataset is based on real 
US economic time series data, but the uncertainty we added is hypothetical 
and included for illustrative purposes.
}
\keyword{datasets}
