% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uncertain_faithfuld}
\alias{uncertain_faithfuld}
\title{2d density estimate of Old Faithful data with uncertainty}
\format{
A data frame with 5,625 observations and 3 variables:
\describe{
  \item{eruptions}{Eruption time in mins}
  \item{waiting}{Waiting time to next eruption in mins}
  \item{density}{A 2d density estimate that is normally distributed with a low variance}
  \item{density2}{A 2d density estimate that is normally distributed with a high variance}
}
}
\description{
A 2d density estimate of the waiting and eruptions variables data
\link{faithful}. Unlike other uncertain datasets, the only uncertain variable is
density. Since this is based on a model, it wouldn't make sense for erruptions or
waiting to be represented as random variables.
}
