% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uncertain_mpg}
\alias{uncertain_mpg}
\title{An uncertain version of the MPG data from `ggplot2`}
\format{
A data frame with 234 rows and 11 variables:
\describe{
  \item{manufacturer}{manufacturer, as a categorical random variable}
  \item{model}{model name as a categorical random variable}
  \item{displ}{engine displacement, as a uniform random variable to represent bounded data}
  \item{year}{year of manufacture, as a sample of possible years}
  \item{cyl}{number of cylinders, as a categorical random variable}
  \item{trans}{type of transmission, as a categorical random variable}
  \item{drv}{the type of drive train, as a categorical random variable }
  \item{cty}{city miles per gallon, as a discrete random variable}
  \item{hwy}{highway miles per gallon, as a discrete random variable}
  \item{fl}{fuel type, as a categorical random variable}
  \item{class}{"type" of car, as a categorical random variable}
}
}
\description{
This dataset is based on the Fuel economy data from 1999 to 2008 from `ggplot2`, 
but every value is represented by a distribution. Every variable in the data set
is represetned by a categorical, discrete, or continuous random variable. The
original MPG dataset in ggplot is a real a subset of the fuel economy data from 
the EPA, but the uncertainty is hypothetical uncertainty for each data type, added
by us for illustrative purposes.
}
