% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uncertain_mtcars}
\alias{uncertain_mtcars}
\title{An uncertain version of the mtcars data from base R `datasets`}
\format{
A data frame with 32 observations and 11 variables:
\describe{
  \item{mpg}{Uniform random variable - Miles/(US) gallon as}
  \item{cyl}{Categorical random variable - Number of cylinders}
  \item{disp}{Uniform random variable - Displacement (cu.in.)}
  \item{hp}{Normal random variable - Gross horsepower}
  \item{drat}{Uniform random variable - Rear axle ratio}
  \item{wt}{Uniform random variable - Weight (1000 lbs)}
  \item{qsec}{Uniform random variable - 1/4 mile time}
  \item{vs}{Bernouli random variable - Engine (0 = V-shaped, 1 = straight)}
  \item{am}{Bernouli random variable - Transmission (0 = automatic, 1 = manual)}
  \item{gear}{Categorical random variable - Number of forward gears}
  \item{carb}{Categorical random variable- Number of carburetors}
}
}
\description{
This dataset is identical to the mtcars data, except that every  variable in the data set
is represented by a categorical, discrete, or continuous random variable. The
original `mtcars` dataset in datasets is based on real data extracted from the 1974 Motor Trend US
magazine, but the uncertainty we added is hypothetical and included for illustrative purposes.
}
