% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar.R
\name{bivariate}
\alias{bivariate}
\alias{is_bivariate}
\title{`bivariate` vector}
\usage{
bivariate(x = distributional::new_dist())

is_bivariate(x)
}
\arguments{
\item{x}{* For `percent()`: a distribution from the `distributional` package
* For `is_percent()`: An object to test.}
}
\value{
An S3 vector of class `ggdibbler_bivariate`.
}
\description{
This creates a bivariate variable that allows us to represent
the mean and variance of a distribution as a single variable.
}
\examples{
bivariate(toy_temp_dist$temp_dist)
}
