% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-sample.R
\name{geom_point_sample}
\alias{geom_point_sample}
\title{Visualise Points Uncertainty}
\usage{
geom_point_sample(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  times = 30,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{times}{A parameter used to control the number of values sampled from each distribution.}
}
\value{
A ggplot2 geom representing a point_sample which can be added to a ggplot object
}
\description{
Identical to geom_point, except that it will accept a distribution in place of any of the usual aesthetics.
}
