% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-sample.R
\name{geom_point_sample}
\alias{geom_point_sample}
\title{Visualise Points Uncertainty}
\usage{
geom_point_sample()
}
\value{
A ggplot2 geom representing a sf_sample which can be added to a ggplot object
}
\description{
Identical to geom_point, except that it will accept a distribution in place of any of the usual aesthetics.
}
