% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sample-point.R
\name{geom_sample_point}
\alias{geom_sample_point}
\title{Visualise Points Uncertainty}
\usage{
geom_sample_point(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\value{
A ggplot2 geom representing a sf_sample which can be added to a ggplot object
}
\description{
Identical to geom_point, except that it will accept a distribution in place of any of the usual aesthetics.
}
