% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_dodge_identity.R
\docType{data}
\name{position_dodge_identity}
\alias{position_dodge_identity}
\alias{PositionDodgeIdentity}
\title{TITLE}
\usage{
position_dodge_identity(
  width = NULL,
  preserve = "total",
  orientation = "x",
  reverse = FALSE
)
}
\description{
DESCRIPTION
}
\examples{
print("x")
}
\keyword{datasets}
