% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodge-nested.R
\docType{data}
\name{position_dodge_dodge}
\alias{position_dodge_dodge}
\alias{position_dodge_identity}
\alias{PositionDodgeIdentity}
\alias{PositionDodgeDodge}
\title{Nested dodge positions}
\usage{
position_dodge_dodge(
  vjust = 1,
  reverse = FALSE,
  width = NULL,
  preserve = "single",
  orientation = "x"
)

position_dodge_identity(
  width = NULL,
  preserve = "single",
  orientation = "x",
  reverse = FALSE
)
}
\arguments{
\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}

\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the \code{"total"} width of all elements
at a position, or the width of a \code{"single"} element?}

\item{orientation}{Fallback orientation when the layer or the data does not
indicate an explicit orientation, like \code{geom_point()}. Can be \code{"x"}
(default) or \code{"y"}.}
}
\description{
These functions use nested positioning for distributional data, where the
original plot has a dodged position. This allows you to set different 
position adjustments for the "main" and "distribution" parts of your plot.
}
\examples{
# Standard ggplots often have a position adjustment to fix overplotting
# plot with dodged positions
ggplot(mpg, aes(class)) + 
  geom_bar(aes(fill = drv), 
           position = position_dodge(preserve = "single"))

# but when we use this in ggdibbler, it can not work the way we expect
# normal dodge
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "dodge")

# nested positions allows us to differentiate which postion adjustments
# are used for the plot groups vs the distribution samples

ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "dodge_identity", alpha=0.2)

# using postion_dodge nests the original plot group inside the distribtion 
# position dodge_dodge does the opposite nesting
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "dodge_dodge")
  
}
\keyword{datasets}
