% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-stack-nested.R
\docType{data}
\name{position_stack_identity}
\alias{position_stack_identity}
\alias{PositionStackIdentity}
\alias{position_stack_dodge}
\alias{PositionStackDodge}
\title{Nested stack positions}
\usage{
position_stack_identity(vjust = 1, reverse = FALSE)

position_stack_dodge(
  vjust = 1,
  reverse = FALSE,
  width = NULL,
  preserve = "single",
  orientation = "x"
)
}
\description{
These functions use nested positioning for distributional data, where the
original plot has a stacked position. This allows you to set different 
position adjustments for the "main" and "distribution" parts of your plot.
}
\examples{
# Standard ggplots often have a position adjustment to fix overplotting
# plot with overplotting
ggplot(mpg, aes(class)) + 
  geom_bar(aes(fill = drv), alpha=0.5,
           position = "stack")

# but when we use this in ggdibbler, it can not work the way we expect
# normal stack warps the scale and doesn't communicate useful info
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "stack")

# nested positions allows us to differentiate which postion adjustments
# are used for the plot groups vs the distribution samples
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "stack_identity", alpha=0.2)
ggplot(uncertain_mpg, aes(class)) + 
  geom_bar_sample(aes(fill = drv), position = "stack_dodge")
}
\keyword{datasets}
