% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-expand.R
\name{sample_expand}
\alias{sample_expand}
\alias{dibble_to_tibble}
\title{Simulate outcomes from dibble to make a tibble}
\usage{
sample_expand(data, times = 10, seed = NULL)

dibble_to_tibble(data, params)
}
\arguments{
\item{data}{Distribution dataset to expand into samples}

\item{times}{A parameter used to control the number of values sampled from 
each distribution.}

\item{seed}{Set the seed for the layers random draw, allows you to get the
same draw from repeated sample_expand calls}

\item{params}{the params argument for the stat function
sample_expand(uncertain_mpg, times=10)}
}
\value{
A data frame of resampled values from the input distributions
}
\description{
Simulates outcomes from all distributions in the dataset to make an "expanded" data
set that can be intepreted by ggplot2. This can be used to debug ggdibbler plots, or
used to make an uncertainty visualisation for a geom that doesn't exist. If
(and only if) you are implementing a ggdibbler version of a ggplot stat
extension, you should use dibble_to_tibble instead.
}
