% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_type.R
\name{scale_type.distribution}
\alias{scale_type.distribution}
\title{Sets scale for distributions}
\usage{
\method{scale_type}{distribution}(x)
}
\arguments{
\item{x}{value being scaled}
}
\value{
A character vector of scale types. The scale type is the ggplot scale type of the outcome of the distribution.
}
\description{
Generates a single value from the distribution and uses it to set the default ggplot scale. The scale can be changed later in the ggplot by using any scale_* function
}
