% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_identity_subdivide.R
\name{subdivide}
\alias{subdivide}
\title{Internal function for subdividing geometry grid}
\usage{
subdivide(geometry, d)
}
\description{
Internal function for subdividing geometry grid
}
\keyword{internal}
