% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uncertain_faithful}
\alias{uncertain_faithful}
\title{Old Faithful data with uncertainty}
\format{
A data frame:
\describe{
  \item{eruptions}{Eruption time in mins}
  \item{waiting}{Waiting time to next eruption in mins}
}
}
\description{
The old faithful data from the datasets package but with added 
uncertainty.
}
